get rid of PhantomData
Since our types are already zero-sized and/or uninstantiable, we can as
well get rid of the PhantomData and use the tuples directly.

Doing so drops 3MB and 3/100 seconds from the build profile, which is
inconsequential, but since it makes the code easier, I'll suggest it
anyway.

Not sure if this is a breaking change, so I'll let you update the
version number.