Fix const-generics feature on 16-bit targets (#224)
Currently, compiling typenum with the const-generics feature on 16-bit
targets causes an error due to conflicting implementations of the ToUInt
trait for Const<0> as the generated code assumes that the pointer width
is at least 32.  This patch adds support for 16-bit targets to the
generated code.

The problem can be reproduced with:
    RUSTFLAGS="-C target-cpu=atmega328p" cargo +nightly build -Z build-std=core --target=avr-none